+++ ced/src/main/java/cnuphys/ced/noise/NoiseManager.java
@@ -102,10 +102,10 @@ public class NoiseManager implements IClasIoEventListener {
 		int hitCount = DC.hitCount();
 		
 		if (hitCount > 0) {
-			int sector[] = DC.sector();
-			int superlayer[] = DC.superlayer();
-			int layer[] = DC.layer();
-			int wire[] = DC.wire();
+			byte sector[] = DC.sector();
+			byte superlayer[] = DC.superlayer();
+			byte layer[] = DC.layer();
+			short wire[] = DC.wire();
 			
 			noisePackage.findNoise(sector,
 					superlayer, layer,
